+++ client/src/index.ts
@@ -15,6 +15,7 @@ import * as path from "path";
 import CreateTaskCommand = require("./commands/createTask");
 import CreateScriptCommand = require("./commands/createScript");
 import CreateREADMECommand = require("./commands/createREADME");
+import CreateMasterCommand = require("./commands/createGitHubActions");
 
 export function activate(context: ExtensionContext) {
   let serverModule = context.asAbsolutePath(
@@ -90,21 +91,28 @@ export function activate(context: ExtensionContext) {
   context.subscriptions.push(disposable);
 	
   // Register commands
-  // create project from template
   let createTask = commands.registerCommand(
     "extension.createTask",
     CreateTaskCommand.run.bind(undefined)
   );
 	context.subscriptions.push(createTask);
+
 	let createScript = commands.registerCommand(
 		"extension.createScript",
 		CreateScriptCommand.run.bind(undefined)
 	)
   context.subscriptions.push(createScript);
+
 	let createREADME = commands.registerCommand(
 		"extension.createREADME",
 		CreateREADMECommand.run.bind(undefined)
 	)
   context.subscriptions.push(createREADME);
+
+	let createMaster = commands.registerCommand(
+		"extension.createMaster",
+		CreateMasterCommand.run.bind(undefined)
+	)
+  context.subscriptions.push(createMaster);
 	
 }